requirements/requirements.txt
langchain==0.2.1
langchain_community==0.2.1
langchain_groq==0.1.4
pandas==2.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
streamlit==1.35.0
sentence-transformers
openpyxl == 3.1.4
pymilvus  == 2.4.3
mistralai == 0.4.1